feat(isa): 1 字节寄存器 ISA 夹具 demo8_v12 + 反回潮守卫 + 文档（P4/P5/P6）
夹具（isa/demo8_v12.toml + arch/demo8_v12.rs）：**唯一 `[reg.gpr1]` 组**的
1 字节寄存器 ISA —— 宽度元数据全 = 1（addr_width / slot_bytes /
value_gpr_width / fp_overhead_bytes = 1）、`default_opsize = 8`（位），
栈对齐 1。历史实现在这个文件上会静默退化（主 GPR 锚点 GPR(8)/GPR(4) 都不存在
⇒ 名字表为空 / 构造不存在的类）。

生成期在这一夹具上又暴露两处残余写死，一并修掉：
- validate_abi 要求 `[abi].stack_align` 是"8 的倍数" → 改为 > 0（栈对齐是
  ISA 数据）；`stack_arg_shadow` 的倍数校验改为按 `[meta].slot_bytes`。
- `@move_args` 整数收参 / 栈参数收参 / by-ref 宽向量 load 基址仍是
  `RegClass::GPR64` 字面量 → `__DEFAULT_GPR_CLASS` / `__ADDR_CLASS`。

新增测试：
- tests/demo8_v12_tests.rs（7 个）：元数据派生（GPR(1)、1 字节槽、帧开销 1、
  sp/fp/scratch 名字解析成功、allocatable = A0..A3）、值池门（i8 Some / i64 None）、
  编码布局、汇编→编码→解码→反汇编文本往返、多态 mov 分发、宿主编译 i8 函数
  （20 字节机器码反汇编 = mov A3,A0 / mov A2,A1 / add A1,A3,A2 / mov A0,A1 / ret）、
  i64 编译期拒绝（点名类型与值池宽度）。
- tests/no_hardcoded_widths.rs × 2（forge-dsl / forge-codegen）：扫描 `src/**`
  生产代码（跳过 `#[cfg(test)]`）禁止写死类字面量；白名单带理由**且条目必须被
  命中**（防白名单腐烂）。开发中它已实际拦下 3 处真实残留。

文档：docs/reference/isa-dsl.md 新增「宽度元数据」节（键/派生规则/生成的
访问器/fail-closed 契约/最小示例 + "指令字宽是另一条轴"的边界说明）、
CLAUDE.md Testing Notes 增一段、CHANGELOG.md 记 Added 条目（含实测数字）。

门禁实测（2026-09-12）：
- cargo fmt --all -- --check 干净；clippy -D warnings 0 告警；
- cargo test --workspace --exclude forge-rustc --exclude cargo-forge
  -j 2 -- --test-threads=1：1320 passed / 0 failed；
- x86 jit matrix：pass=195 skip=3 fail=0（FORGE_JIT_EVENTS 核对）；
- RUSTDOCFLAGS=-D warnings cargo doc --no-deps --all-features 干净；
- markdownlint-cli2（isa-dsl.md / CLAUDE.md / CHANGELOG.md）0 error。

+++ crates/backend/forge-codegen/src/pipeline/lowering.rs
@@ -251,7 +251,7 @@ impl<I: crate::machine::inst::MachineInst + 'static> CompileState<I> {
             }
         }
         for v in iadd_stack_offsets {
-            let depth = -v + slot_unit as i64;
+            let depth = -v + slot_unit;
             stackaddr_depth = stackaddr_depth.max(depth);
             // 与主循环 StackAddr immediate 的处理一致：负偏移槽深计入 locals
             // 帧需求（否则 spill 槽从过浅位置分配覆盖局部变量）。
@@ -263,7 +263,7 @@ impl<I: crate::machine::inst::MachineInst + 'static> CompileState<I> {
                 match inst.opcode {
                     Opcode::StackAddr => {
                         if let Some(Immediate::Int(v)) = inst.immediates.first() {
-                            let depth = if *v >= 0 { 0 } else { -*v + slot_unit as i64 };
+                            let depth = if *v >= 0 { 0 } else { -*v + slot_unit };
                             stackaddr_depth = stackaddr_depth.max(depth);
                         }
                     }
